# openapi.yaml
openapi: 3.0.0
info:
  title: ShopBuilder API
  version: 2.0.0
  description: >-

    ## ShopBuilder — Enterprise Multi-Tenant E-Commerce Platform


    A scalable multi-tenant backend supporting multiple merchants, stores, and customers within isolated tenant
    namespaces.


    ### Role-Based Access Control (RBAC)


    | Role | Description |

    |------|-------------|

    | `SUPER_ADMIN` | Full platform access — manages tenants, users, everything |

    | `PLATFORM_ADMIN` | Manages merchants and platform configuration |

    | `MERCHANT_OWNER` | Manages own stores, products, and orders |

    | `STORE_MANAGER` | Manages inventory and orders for assigned store |

    | `CUSTOMER` | Places and views own orders |


    ### Authentication

    Include a Bearer token on all protected endpoints:

    ```

    Authorization: Bearer <access_token>

    ```


    Access tokens expire in **15 minutes**. Use `POST /auth/refresh` with your refresh token to get a new one.


    ### Multi-Tenancy

    All business data is scoped to a `tenantId`. It is embedded in the JWT and automatically applied.

    Admins can override `tenantId` via query parameter.
          
  contact:
    name: ShopBuilder API Support
    email: api@shopbuilder.io
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
  schemas:
...
    Store:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Downtown Flagship
        description:
          type: string
          nullable: true
        merchantId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    Variant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          example: prod-uuid-S-Black-Cotton
        barcode:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        material:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        stock:
          type: integer
          example: 100
        reservedStock:
          type: integer
          example: 15
        productId:
          type: string
          format: uuid
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
          example: ORD-20240115-83421
        status:
          type: string
          enum:
            - PENDING
            - PAID
            - PACKED
            - SHIPPED
            - DELIVERED
            - CANCELLED
            - REFUNDED
        totalAmount:
          type: number
          example: 59.98
        notes:
          type: string
          nullable: true
        storeId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cartId:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
        quantity:
          type: integer
          example: 2
        variant:
          $ref: '#/components/schemas/Variant'
    Cart:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        subtotal:
          type: number
          example: 89.97
        store:
          $ref: '#/components/schemas/Store'
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
security:
  - bearerAuth: []
paths:
  /admin/stats:
    get:
      tags:
        - Admin
      summary: Platform-wide statistics
      description: Returns total counts for users, tenants, merchants, stores, products, and orders — grouped by status/role.
      responses:
        '200':
          description: Platform stats
          content:
            application/json:
              example:
                totals:
                  users: 42
                  tenants: 3
                  merchants: 8
                  stores: 12
                  products: 156
                  orders: 319
                ordersByStatus:
                  PENDING: 24
                  PAID: 101
                  DELIVERED: 187
                usersByRole:
                  CUSTOMER: 35
                  MERCHANT_OWNER: 5
        '403':
          description: Forbidden
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all platform users
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum:
              - SUPER_ADMIN
              - PLATFORM_ADMIN
              - MERCHANT_OWNER
              - STORE_MANAGER
              - CUSTOMER
        - in: query
          name: tenantId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated user list
  /admin/users/{id}:
    get:
      tags:
        - Admin
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
        '404':
          description: User not found
    delete:
      tags:
        - Admin
      summary: Delete a user (SUPER_ADMIN only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted
        '404':
          description: User not found
  /admin/users/{id}/role:
    patch:
      tags:
        - Admin
      summary: Update a user's role and/or tenant assignment (SUPER_ADMIN only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum:
                    - SUPER_ADMIN
                    - PLATFORM_ADMIN
                    - MERCHANT_OWNER
                    - STORE_MANAGER
                    - CUSTOMER
                  example: MERCHANT_OWNER
                tenantId:
                  type: string
                  format: uuid
                  nullable: true
                  description: Assign or remove tenant association
      responses:
        '200':
          description: User role updated
        '400':
          description: Invalid role
        '404':
          description: User not found
  /admin/tenants:
    get:
      tags:
        - Admin
      summary: List all tenants with aggregate counts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated tenant list
    post:
      tags:
        - Admin
      summary: Create a new tenant (isolates a merchant ecosystem)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
              properties:
                name:
                  type: string
                  example: Acme Commerce
                slug:
                  type: string
                  example: acme-commerce
                  description: URL-safe unique identifier
      responses:
        '201':
          description: Tenant created
        '409':
          description: Slug already taken
  /admin/tenants/{id}:
    get:
      tags:
        - Admin
      summary: Get a tenant by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant detail
        '404':
          description: Tenant not found
...
  /cart:
    get:
      tags:
        - Cart
      summary: Get the current user's cart for a store
      parameters:
        - in: query
          name: storeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cart with items and subtotal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: storeId is required
        '401':
          description: Unauthorized
    delete:
      tags:
        - Cart
      summary: Clear all items from the cart
      parameters:
        - in: query
          name: storeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cart cleared
        '401':
          description: Unauthorized
  /cart/items:
    post:
      tags:
        - Cart
      summary: Add an item to the cart
      description: Creates the cart if it doesn't exist. Adds to existing quantity if the variant is already in the cart.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - storeId
                - variantId
              properties:
                storeId:
                  type: string
                  format: uuid
                variantId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  default: 1
      responses:
        '201':
          description: Item added or quantity updated
        '400':
          description: Insufficient stock or missing fields
        '401':
          description: Unauthorized
        '404':
          description: Store or variant not found
  /cart/items/{variantId}:
    patch:
      tags:
        - Cart
      summary: Update item quantity in cart
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - storeId
                - quantity
              properties:
                storeId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Item quantity updated
        '400':
          description: Insufficient stock or invalid quantity
        '401':
          description: Unauthorized
        '404':
          description: Cart or item not found
    delete:
      tags:
        - Cart
      summary: Remove an item from the cart
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: storeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item removed
        '401':
          description: Unauthorized
        '404':
          description: Item not found
  /cart/checkout:
    post:
      tags:
        - Cart
      summary: Convert cart to order (checkout)
      description: |
        Creates an order from the current cart, reserves stock, clears the cart,
        and sends an order confirmation email. The order starts in PENDING status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - storeId
              properties:
                storeId:
                  type: string
                  format: uuid
                notes:
                  type: string
                  example: Please leave at the door
      responses:
        '201':
          description: Order created — confirmation email queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Empty cart or insufficient stock
        '401':
          description: Unauthorized
        '404':
          description: Store not found
...